sci-libs/arrayfire: Add nonfree use flag
Package-Manager: portage-2.2.20.1

+++ sci-libs/arrayfire/arrayfire-9999.ebuild
@@ -4,7 +4,7 @@
 
 EAPI=5
 
-inherit cmake-utils git-r3
+inherit cmake-utils git-r3 multilib
 
 GTEST_PV="1.7.0"
 
@@ -14,9 +14,10 @@ EGIT_REPO_URI="https://github.com/${PN}/${PN}.git git://github.com/${PN}/${PN}.g
 SRC_URI="test? ( https://googletest.googlecode.com/files/gtest-${GTEST_PV}.zip )"
 KEYWORDS=""
 
-LICENSE="BSD"
+LICENSE="BSD
+	nonfree? ( OpenSIFT )"
 SLOT="0"
-IUSE="+examples +cpu cuda opencl test graphics"
+IUSE="+examples +cpu cuda nonfree opencl test graphics"
 
 RDEPEND="
 	>=sys-devel/gcc-4.7:*
@@ -65,7 +66,9 @@ pkg_pretend() {
 src_unpack() {
 	git-r3_src_unpack
 
-	find "${WORKDIR}" -name "*_nonfree*" -delete || die
+	if ! use nonfree; then
+		find "${WORKDIR}" -name "*_nonfree*" -delete || die
+	fi
 
 	if use test; then
 		mkdir -p "${BUILD_DIR}"/third_party/src/ || die
@@ -89,7 +92,7 @@ src_configure() {
 	   $(cmake-utils_use_build examples EXAMPLES)
 	   $(cmake-utils_use_build test TEST)
 	   $(cmake-utils_use_build graphics GRAPHICS)
-	   -DBUILD_NONFREE=OFF
+	   $(cmake-utils_use_build nonfree NONFREE)
 	   -DUSE_SYSTEM_BOOST_COMPUTE=ON
 	   -DUSE_SYSTEM_CLBLAS=ON
 	   -DUSE_SYSTEM_CLFFT=ON
